Loadng Works! (Kinda-sorta)
The good news: It works. The things that need to be done are done in roughly the right order to load a project.
The bad news: This is a prototype. Not elegant/battle tested enough to make it into production. I should diff this w/ the non-Loading Version to bring this prototype within the norms of production code.